requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
drf-spectacular>=0.15.1,<0.16
pyserial>=3.5,<3.6
requests>=2.28.2,<2.29